+++ aws-test/tests/aws_ec2_transit_gateway_route/test-list-query.sql
@@ -1,0 +1,3 @@
+select transit_gateway_route_table_id, destination_cidr_block, state, type
+from aws_ec2_transit_gateway_route
+where transit_gateway_route_table_id = '{{ output.transit_gateway_rtb_id.value }}';