+++ aws-test/tests/aws_organizations_account/test-list-query.sql
@@ -1,0 +1,3 @@
+select id, arn, email, name, status
+from aws.aws_organizations_account
+where arn = '{{ output.resource_arn.value }}';