+++ aws/table_aws_ec2_instance.go
@@ -390,6 +390,10 @@ func listEc2Instance(ctx context.Context, d *plugin.QueryData, _ *plugin.Hydrate
 	},
 	)
 
+	if err != nil {
+		plugin.Logger(ctx).Error("listEc2Instance", "DescribeInstancesPages_error", err)
+	}
+
 	return nil, err
 }
 
@@ -413,6 +417,7 @@ func getEc2Instance(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateD
 
 	op, err := svc.DescribeInstances(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getEc2Instance", "DescribeInstances_error", err)
 		return nil, err
 	}
 
@@ -432,6 +437,7 @@ func getEc2InstanceARN(ctx context.Context, d *plugin.QueryData, h *plugin.Hydra
 	getCommonColumnsCached := plugin.HydrateFunc(getCommonColumns).WithCache()
 	commonData, err := getCommonColumnsCached(ctx, d, h)
 	if err != nil {
+		plugin.Logger(ctx).Error("getEc2InstanceARN", "getCommonColumnsCached_error", err)
 		return nil, err
 	}
 	commonColumnData := commonData.(*awsCommonColumnData)
@@ -459,6 +465,7 @@ func getInstanceDisableAPITerminationData(ctx context.Context, d *plugin.QueryDa
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceDisableAPITerminationData", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -483,6 +490,7 @@ func getInstanceInitiatedShutdownBehavior(ctx context.Context, d *plugin.QueryDa
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceInitiatedShutdownBehavior", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -507,6 +515,7 @@ func getInstanceKernelID(ctx context.Context, d *plugin.QueryData, h *plugin.Hyd
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceKernelID", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -531,6 +540,7 @@ func getInstanceRAMDiskID(ctx context.Context, d *plugin.QueryData, h *plugin.Hy
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceRAMDiskID", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -555,6 +565,7 @@ func getInstanceSriovNetSupport(ctx context.Context, d *plugin.QueryData, h *plu
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceSriovNetSupport", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -579,6 +590,7 @@ func getInstanceUserData(ctx context.Context, d *plugin.QueryData, h *plugin.Hyd
 
 	instanceData, err := svc.DescribeInstanceAttribute(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceUserData", "DescribeInstanceAttribute_error", err)
 		return nil, err
 	}
 
@@ -603,6 +615,7 @@ func getInstanceStatus(ctx context.Context, d *plugin.QueryData, h *plugin.Hydra
 
 	instanceData, err := svc.DescribeInstanceStatus(params)
 	if err != nil {
+		plugin.Logger(ctx).Error("getInstanceStatus", "DescribeInstanceStatus_error", err)
 		return nil, err
 	}